feat: close order adding fee to db

+++ src/models/fee_model.rs
@@ -1,5 +1,5 @@
 use chrono::{DateTime, Utc};
-use rust_decimal::Decimal;
+// use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 // use utoipa::ToSchema;
 
@@ -15,5 +15,5 @@ pub struct Fee {
 pub struct CreateFeeRequest {
     pub user_id: String,
     pub agent_id: String,
-    pub amount: Decimal,
+    pub amount: f64,
 }
